feat: adiciona exemplos context
