+++ src/components/Consumidor/Consumidor.js
@@ -1,0 +1,7 @@
+import { useAutenticacao } from '../../contexts/autenticacao';
+
+export const Consumidor = () => {
+  const { nome } = useAutenticacao();
+
+  return <p>{`Consumidor autenticado: ${nome}`}</p>;
+};
