+++ src/components/Consumidor/index.js
@@ -1,0 +1,2 @@
+export * from './Consumidor';
+export * from './Consumidor2';
