+++ src/components/Header/Header.js
@@ -1,5 +1,19 @@
-import styles from "./Header.module.css";
+import { usePokemon } from '../../contexts/pokemon';
+import styles from './Header.module.css';
 
 export const Header = () => {
-  return <div className={styles.container}>Header</div>;
+  const { setTermoBusca } = usePokemon();
+
+  return (
+    <div className={styles.container}>
+      <input
+        onChange={(event) => {
+          setTermoBusca(event.target.value);
+        }}
+        className={styles.searchInput}
+        type="text"
+        placeholder="Digite o nome do pokemon"
+      />
+    </div>
+  );
 };
