+++ src/components/Header/Header.module.css
@@ -9,3 +9,9 @@
   justify-content: center;
   align-items: center;
 }
+
+.searchInput {
+  padding: 1rem 2rem;
+  font-size: 1.25rem;
+  margin: 1rem 0;
+}
