+++ src/components/Pokemons/Pokemons.module.css
@@ -24,12 +24,6 @@
   gap: 10px;
 }
 
-.searchInput {
-  padding: 1rem 2rem;
-  font-size: 1.25rem;
-  margin: 1rem 0;
-}
-
 .pagination {
   display: flex;
   justify-content: space-between;
