+++ src/contexts/autenticacao/AutenticacaoContext.js
@@ -1,0 +1,3 @@
+import { createContext } from 'react';
+
+export const AutenticacaoContext = createContext({ autenticado: true, nome: 'Maria' });
