+++ src/contexts/autenticacao/AutenticacaoProvider.js
@@ -1,0 +1,8 @@
+import { useState } from 'react';
+import { AutenticacaoContext } from './AutenticacaoContext';
+
+export const AutenticacaoProvider = ({ children }) => {
+  const [nome, setNome] = useState('Thais');
+
+  return <AutenticacaoContext.Provider value={{ nome, setNome }}>{children}</AutenticacaoContext.Provider>;
+};
