+++ src/contexts/autenticacao/index.js
@@ -1,0 +1,3 @@
+export * from './AutenticacaoContext';
+export * from './useAutenticacao';
+export * from './AutenticacaoProvider';
