+++ src/contexts/autenticacao/useAutenticacao.js
@@ -1,0 +1,4 @@
+import { useContext } from 'react';
+import { AutenticacaoContext } from './AutenticacaoContext';
+
+export const useAutenticacao = () => useContext(AutenticacaoContext);
