+++ src/contexts/pokemon/usePokemon.js
@@ -1,0 +1,4 @@
+import { useContext } from 'react';
+import { PokemonContext } from './PokemonContext';
+
+export const usePokemon = () => useContext(PokemonContext);
